+++ packages/terra-application-layout/src/terra-dev-site/doc/example/ExampleApplication.jsx
@@ -243,6 +243,13 @@ class ExampleApplication extends React.Component {
       title: 'Custom Checkbox Item',
       isSelectable: true,
       isSelected: checkboxItemEnabled,
+      isReadOnly: false,
+      parentKey: Utils.utilityHelpers.defaultKeys.MENU,
+    }, {
+      key: 'readonly-item',
+      title: 'Custom ReadOnly Item',
+      isSelectable: true,
+      isReadOnly: true,
       parentKey: Utils.utilityHelpers.defaultKeys.MENU,
     }, {
       key: 'additional-3',