+++ packages/terra-hookshot/src/terra-dev-site/doc/example/HookshotExample.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Button from 'terra-button';
-import NumberField from 'terra-form/lib/NumberField';
+import InputField from 'terra-form-input/lib/InputField';
+import SelectField from 'terra-form-select/lib/SelectField';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import Hookshot from 'terra-hookshot/lib/Hookshot';
 
@@ -19,7 +20,7 @@ const ATTACHMENT_POSITIONS = [
 const generateOptions = values => (
   values.map((currentValue, index) => {
     const keyValue = index;
-    return <option key={keyValue} value={currentValue}>{currentValue}</option>;
+    return <SelectField.Option key={keyValue} value={currentValue} display={currentValue} />;
   })
 );
 
@@ -44,14 +45,17 @@ const attachmentValues = (attachment) => {
   return { vertical: 'bottom', horizontal: 'center' };
 };
 
+const getId = name => `${name}Example`;
+
 class HookshotStandard extends React.Component {
   constructor(props) {
     super(props);
     this.handleButtonClick = this.handleButtonClick.bind(this);
     this.handleRequestClose = this.handleRequestClose.bind(this);
-    this.handleSelectChange = this.handleSelectChange.bind(this);
+    this.handleAttachementBehaviorChange = this.handleAttachementBehaviorChange.bind(this);
+    this.handleContentAttachmentChange = this.handleContentAttachmentChange.bind(this);
+    this.handleTargetAttachmentChange = this.handleTargetAttachmentChange.bind(this);
     this.handleInputChange = this.handleInputChange.bind(this);
-    this.getId = this.getId.bind(this);
     this.state = {
       isOpen: false,
       hookshotContentAttachment: ATTACHMENT_POSITIONS[1],
@@ -59,10 +63,6 @@ class HookshotStandard extends React.Component {
       hookshotAttachmentBehavior: Hookshot.attachmentBehaviors[0],
       hookshotAttachmentMargin: 0,
     };
-  }
-
-  getId(name) {
-    return name + this.state.id;
   }
 
   handleButtonClick() {
@@ -73,12 +73,21 @@ class HookshotStandard extends React.Component {
     this.setState({ isOpen: false });
   }
 
-  handleSelectChange(event) {
-    this.setState({ [event.target.name]: event.target.value });
+  handleAttachementBehaviorChange(value) {
+    this.setState({ hookshotAttachmentBehavior: value });
+  }
+
+  handleContentAttachmentChange(value) {
+    this.setState({ hookshotContentAttachment: value });
+  }
+
+  handleTargetAttachmentChange(value) {
+    this.setState({ hookshotTargetAttachment: value });
   }
 
   handleInputChange(event) {
-    this.setState({ [event.target.name]: Number.parseFloat(event.target.value) });
+    const value = Number.parseFloat(event.target.value);
+    this.setState({ [event.target.name]: value });
   }
 
   render() {
@@ -94,35 +103,41 @@ class HookshotStandard extends React.Component {
 
     return (
       <div>
-        <form>
-          <label htmlFor={this.getId('hookshotAttachmentBehavior')}>Attachment Behavior</label>
-          <select id={this.getId('hookshotAttachmentBehavior')} name="hookshotAttachmentBehavior" value={this.state.hookshotAttachmentBehavior} onChange={this.handleSelectChange}>
-            {generateOptions(Hookshot.attachmentBehaviors)}
-          </select>
-          <br />
-          <br />
-          <NumberField
-            label="Attachment Margin in Px"
-            id={this.getId('hookshotAttachmentMargin')}
-            name="hookshotAttachmentMargin"
-            defaultValue={0}
-            value={this.state.hookshotAttachmentMargin}
-            style={{ width: '200px' }}
-            onChange={this.handleInputChange}
-          />
-          <label htmlFor={this.getId('hookshotContentAttachment')}>Content Attachment</label>
-          <select id={this.getId('hookshotContentAttachment')} name="hookshotContentAttachment" value={this.state.hookshotContentAttachment} onChange={this.handleSelectChange}>
-            {generateOptions(ATTACHMENT_POSITIONS)}
-          </select>
-          <br />
-          <br />
-          <label htmlFor={this.getId('hookshotTargetAttachment')}>Target Attachment</label>
-          <select id={this.getId('hookshotTargetAttachment')} name="hookshotTargetAttachment" value={this.state.hookshotTargetAttachment} onChange={this.handleSelectChange}>
-            {generateOptions(ATTACHMENT_POSITIONS)}
-          </select>
-          <br />
-          <br />
-        </form>
+        <SelectField
+          label="Attachment Behavior"
+          selectId={getId('hookshotAttachmentBehavior')}
+          selectAttrs={{ name: 'hookshotAttachmentBehavior' }}
+          value={this.state.hookshotAttachmentBehavior}
+          onChange={this.handleAttachementBehaviorChange}
+        >
+          {generateOptions(Hookshot.attachmentBehaviors)}
+        </SelectField>
+        <InputField
+          label="Attachment Margin in Px"
+          inputId={getId('hookshotAttachmentMargin')}
+          inputAttrs={{ name: 'hookshotAttachmentMargin', type: 'number' }}
+          defaultValue={this.state.hookshotAttachmentMargin}
+          style={{ width: '200px' }}
+          onChange={this.handleInputChange}
+        />
+        <SelectField
+          label="Content Attachment"
+          selectId={getId('hookshotContentAttachment')}
+          selectAttrs={{ name: 'hookshotContentAttachment' }}
+          value={this.state.hookshotContentAttachment}
+          onChange={this.handleContentAttachmentChange}
+        >
+          {generateOptions(ATTACHMENT_POSITIONS)}
+        </SelectField>
+        <SelectField
+          label="Target Attachment"
+          selectId={getId('hookshotTargetAttachment')}
+          selectAttrs={{ name: 'hookshotTargetAttachment' }}
+          value={this.state.hookshotTargetAttachment}
+          onChange={this.handleTargetAttachmentChange}
+        >
+          {generateOptions(ATTACHMENT_POSITIONS)}
+        </SelectField>
         <Hookshot
           attachmentBehavior={this.state.hookshotAttachmentBehavior}
           attachmentMargin={this.state.hookshotAttachmentMargin}