+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -4,6 +4,19 @@ Changelog
 Unreleased
 ----------
 
+1.20.0 - (January 5, 2019)
+------------------
+### Changed
+* Minor dependency update
+
+1.19.0 - (January 2, 2019)
+------------------
+### Fixed
+* Updated Data-menu-item attribute value of Menuitem to Unique key value
+
+### Added
+* Add optional toolbar prop that displays below the side menu action header
+
 1.18.0 - (December 5, 2018)
 ------------------
 ### Changed